pyproject.toml
[tool.poetry]
name = "aws"
version = "0.1.0"
description = "AWS accounts as code"
authors = ["Brodie Blackburn <brodie.blackburn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "^2.27.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
